requirements.txt
asgiref==3.8.1
boto3==1.35.36
botocore==1.35.36
certifi==2024.8.30
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gevent==24.2.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
packaging==24.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
zope.event==5.0
zope.interface==7.0.3
channels==4.1.0
daphne==4.1.2